Fixed external program selection * Original 1.x branch used a text entry widget with a button   next to it, which launched an application chooser dialog * The 2.x branch changed to an aapchooser button, which only partway worked   * The program could be selected and saved into the preferences   * The next time the prefs dialog ran, the default program for image/svg+xml     was displayed on the button
Explanation: the gtk::AppChooser button does not provide a good way to set the
default application programmatically. One must append a custom item and set
that as the default, but there is no interface for matching the custom item to
a real application from the generated list. It could be done by iterating over
the entire list to match against the saved commandline, but this is simpler.